ajout du profile administrateur et des favoris des publications et limitation des delete

+++ back/controllers/publication.js
@@ -1,4 +1,5 @@
 const db = require('../db');
+const fs = require('fs');
 const { decodeToken } = require('../middleware/auth');
 const { getNumberComments } = require('./comment');
 const { getUserLikePublication } = require('./like');
@@ -37,10 +38,75 @@ exports.createPublication = (req, res, next) => {
             });
         }
     );
-    /*
-    */
 }
 
+exports.getMorePopularPublication = async (req, res, next) => {
+    const sql = "SELECT count(publicationId) as nb, publicationId FROM groupomania.comments group by publicationId order by nb DESC;"
+    let results = null;
+    try {
+        results = await db.sequelize.query(sql, {
+            type: db.sequelize.QueryTypes.SELECT
+        });
+
+    } catch (error) {
+        console.warn(error);
+        res.status(500).json("Une erreure est arrivé en essayant de recuperer la liste des publications les plus populaires");
+    }
+    const publications = [];
+    for (r of results) {
+        let p = {};
+
+        await Publication.findOne({
+            where: { id: r.publicationId },
+            raw: true
+        }).then((foundPblication) => {
+            p = foundPblication;
+        });
+        p.popularity = r.nb;
+
+        await getUserById(p.userId).then(u => {
+            p.user = u
+        });
+
+        await getNumberComments(p.id).then(c => {
+            p.comments = c;
+        })
+
+        await getUserLikePublication(p.id).then(l => {
+            p.iLike = l.like;
+            p.iDislike = l.dislike;
+        });
+
+        publications.push(p);
+    }
+    res.status(200).json(publications);
+}
+
+exports.updatePublication = (req, res, next) => {
+    // On récupère les infos envoyés depuis un formulaire
+    // le json de la publication est mis dans une chaine de caractère
+    // on parse le json
+    const publicationObject = req.body;
+    const user_id = decodeToken(req).id;
+
+    Publication.update(
+        {
+            title: publicationObject.title,
+            description: publicationObject.description
+        },
+        { where: { id: req.params.id, userId: user_id }, raw: true }
+    ).then(
+        (publication) => {
+            res.status(200).json(publication);
+        }
+    ).catch(
+        (error) => {
+            res.status(404).json({
+                error
+            });
+        }
+    );
+}
 
 
 exports.getAllPublication = async (req, res, next) => {
@@ -61,6 +127,7 @@ exports.getAllPublication = async (req, res, next) => {
             }
         );
     } catch (error) {
+        console.warn(error);
     }
 
     for (p of publications) {
@@ -112,8 +179,13 @@ exports.getOnePublication = async (req, res, next) => {
 };
 
 exports.deletePublication = (req, res, next) => {
-    Publication.findOne({ where: { id: req.params.id } })
+    const user = decodeToken(req);
+    Publication.findOne({ where: { id: req.params.id }, raw: true })
         .then(p => {
+            if (user.id != p.userId && !user.isAdmin) {
+                res.status(403).json({ message: 'Vous ne pouvez pas supprimer cette publication' });
+                return;
+            }
             const filename = p.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Publication.destroy({ where: { id: req.params.id } })
